Move tests into solver folder
