+++ generate_answers.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-docker compose -f docker-compose-dev.yml run -v $(pwd):/bot bot python -m tests.data.generate_answers $*
+docker compose -f docker-compose-dev.yml run solver python -m solver.tests.generate_answers $*
