+++ run_tests.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-docker compose -f docker-compose-dev.yml run -v $(pwd):/solver solver python -m pytest -s -x -vv -n logical $*
+docker compose -f docker-compose-dev.yml run solver python -m pytest --import-mode=importlib -s -x -vv -n logical $*
