+++ solver/tests/helper.py
@@ -1,0 +1,3 @@
+from pathlib import Path
+
+data_path = Path(__file__).parent / 'data'
